fix: proper TypeScript integration
