--- .npmignore
@@ -1,16 +1,0 @@
-*.md
-rustfmt.toml
-.all-contributorsrc
-.github/
-**/target/
-**/tsbuild/
-synth/migrations/
-synth/testing_harness/
-synth/tests/
-synth/wix/
-docs/
-examples/
-www/
-tools/
-nix/
-*.nix
