+++ node/content/bool.ts
