+++ node/content/object.ts
