+++ node/content/string.ts
